AsBatched() overhaul, basic tests added. C# interface implementing as extensions.
