+++ Composite/Composite.Cs.Tests/Tests.cs
@@ -16,53 +16,6 @@ namespace Composite.Cs.Tests {
         }
 
         [Fact]
-        public void ToBatchedTest () {
-            var inputSequence = Enumerable.Range(1, 10)
-                .Select(x => new Simple{ Number = x, })
-                .AsLimited(6);
-            var batches = C.ToBatched(10, x=>x.Number,  inputSequence)
-                .Take(2)
-                .ToArray();
-
-            Assert.Equal(1, batches[0][0].Number);
-            Assert.Equal(2, batches[0][1].Number);
-            Assert.Equal(3, batches[0][2].Number);
-            Assert.Equal(4, batches[0][3].Number);
-            Assert.Equal(5, batches[1][0].Number);
-            Assert.Equal(6, batches[1][1].Number);
-        }
-
-        [Fact]
-        public void ToPagedTest () {
-            var inputSequence = Enumerable.Range(1, 10).AsLimited(6);
-            var pages = C.ToPaged(2, inputSequence).Take(3).ToArray();
-
-            Assert.Equal(1, pages[0][0]);
-            Assert.Equal(2, pages[0][1]);
-            Assert.Equal(3, pages[1][0]);
-            Assert.Equal(4, pages[1][1]);
-            Assert.Equal(5, pages[2][0]);
-            Assert.Equal(6, pages[2][1]);
-        }
-
-        [Fact]
-        public void ToPartitionedTest () {
-            var inputSequence = Enumerable.Range(1, 10).AsLimited(6);
-            var partitions = C.ToPartitioned(3, inputSequence);
-
-            var arr1 = partitions[0].Take(2).ToArray();
-            var arr2 = partitions[1].Take(2).ToArray();
-            var arr3 = partitions[2].Take(2).ToArray();
-
-            Assert.Equal(1, arr1[0]);
-            Assert.Equal(2, arr1[1]);
-            Assert.Equal(3, arr2[0]);
-            Assert.Equal(4, arr2[1]);
-            Assert.Equal(5, arr3[0]);
-            Assert.Equal(6, arr3[1]);
-        }
-
-        [Fact]
         public void ToForestTest(){
             var obj = C.Composite (new [] {
                 C.Value(new Simple { Number = 2, }),
